Update the deletion behavior of surveys.

+++ app/org/sagebionetworks/bridge/dynamodb/DynamoSurveyDao.java
@@ -77,8 +77,8 @@ public class DynamoSurveyDao implements SurveyDao {
             this.published = true;
             return this;
         }
-        QueryBuilder isNotDeleted() {
-            this.notDeleted = true;
+        QueryBuilder setDeleted(boolean includeDeleted) {
+            this.notDeleted = !includeDeleted;
             return this;
         }
         
@@ -259,7 +259,7 @@ public class DynamoSurveyDao implements SurveyDao {
     }
 
     @Override
-    public Survey publishSurvey(StudyIdentifier study, Survey survey, GuidCreatedOnVersionHolder keys, boolean newSchemaRev) {
+    public Survey publishSurvey(StudyIdentifier study, Survey survey, boolean newSchemaRev) {
         if (survey.isDeleted()) {
             throw new EntityNotFoundException(Survey.class);
         }
@@ -286,9 +286,6 @@ public class DynamoSurveyDao implements SurveyDao {
     @Override
     public Survey updateSurvey(Survey survey) {
         Survey existing = getSurvey(survey, false);
-        if (existing.isDeleted()) {
-            throw new EntityNotFoundException(Survey.class);
-        }
         if (existing.isPublished()) {
             throw new PublishedSurveyException(survey);
         }
@@ -312,9 +309,6 @@ public class DynamoSurveyDao implements SurveyDao {
     @Override
     public Survey versionSurvey(GuidCreatedOnVersionHolder keys) {
         DynamoSurvey existing = (DynamoSurvey)getSurvey(keys, true);
-        if (existing.isDeleted()) {
-            throw new EntityNotFoundException(Survey.class);
-        }
         DynamoSurvey copy = new DynamoSurvey(existing);
         copy.setPublished(false);
         copy.setDeleted(false);
@@ -339,46 +333,50 @@ public class DynamoSurveyDao implements SurveyDao {
 
     @Override
     public void deleteSurveyPermanently(GuidCreatedOnVersionHolder keys) {
-        Survey existing = getSurvey(keys, false);
-        deleteAllElements(existing.getGuid(), existing.getCreatedOn());
-        surveyMapper.delete(existing);
         
-        // Delete the schemas as well, or they accumulate.
-        try {
-            StudyIdentifier studyId = new StudyIdentifierImpl(existing.getStudyIdentifier());
-            uploadSchemaService.deleteUploadSchemaById(studyId, existing.getIdentifier());
-        } catch(EntityNotFoundException e) {
-            // This is OK. Just means this survey wasn't published.
+        Survey existing = new QueryBuilder().setSurvey(keys.getGuid()).setCreatedOn(keys.getCreatedOn())
+                .setSkipElements(true).getOne(false);
+        if (existing != null) {
+            deleteAllElements(existing.getGuid(), existing.getCreatedOn());
+            surveyMapper.delete(existing);
+            
+            // Delete the schemas as well, or they accumulate.
+            try {
+                StudyIdentifier studyId = new StudyIdentifierImpl(existing.getStudyIdentifier());
+                uploadSchemaService.deleteUploadSchemaById(studyId, existing.getIdentifier());
+            } catch(EntityNotFoundException e) {
+                // This is OK. Just means this survey wasn't published.
+            }
         }
     }
 
     @Override
-    public List<Survey> getSurveyAllVersions(StudyIdentifier studyIdentifier, String guid) {
-        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).isNotDeleted().getAll(true);
+    public List<Survey> getSurveyAllVersions(StudyIdentifier studyIdentifier, String guid, boolean includeDeleted) {
+        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).setDeleted(includeDeleted).getAll(true);
     }
     
     @Override
     public Survey getSurveyMostRecentVersion(StudyIdentifier studyIdentifier, String guid) {
-        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).isNotDeleted().getOne(true);
+        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).setDeleted(false).getOne(true);
     }
 
     @Override
     public Survey getSurveyMostRecentlyPublishedVersion(StudyIdentifier studyIdentifier, String guid, boolean includeElements) {
-        return new QueryBuilder().setStudy(studyIdentifier).isPublished().setSurvey(guid).isNotDeleted()
+        return new QueryBuilder().setStudy(studyIdentifier).isPublished().setSurvey(guid).setDeleted(false)
                 .setSkipElements(!includeElements).getOne(true);
     }
     
     // secondary index query (not survey GUID) 
     @Override
-    public List<Survey> getAllSurveysMostRecentlyPublishedVersion(StudyIdentifier studyIdentifier) {
-        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).isPublished().isNotDeleted().getAll(false);
+    public List<Survey> getAllSurveysMostRecentlyPublishedVersion(StudyIdentifier studyIdentifier, boolean includeDeleted) {
+        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).isPublished().setDeleted(includeDeleted).getAll(false);
         return findMostRecentVersions(surveys);
     }
     
     // secondary index query (not survey GUID)
     @Override
-    public List<Survey> getAllSurveysMostRecentVersion(StudyIdentifier studyIdentifier) {
-        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).isNotDeleted().getAll(false);
+    public List<Survey> getAllSurveysMostRecentVersion(StudyIdentifier studyIdentifier, boolean includeDeleted) {
+        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).setDeleted(includeDeleted).getAll(false);
         return findMostRecentVersions(surveys);
     }
